requirements.txt
# Core Dependencies
pandas==2.2.0
numpy==1.26.0
scikit-learn==1.4.0
sqlalchemy==2.0.25

# Web Scraping & APIs
requests==2.31.0
beautifulsoup4==4.12.3
selenium==4.18.1
scrapy==2.11.0
tweepy==4.14.0  # Twitter API
facebook-sdk==3.1.0

# Natural Language Processing
spacy==3.7.2
nltk==3.8.1
transformers==4.37.2
sentence-transformers==2.3.1

# IRS Data Processing
xmltodict==0.13.0
python-irspy==0.3.0  # IRS 990 data

# Geocoding & Mapping
geopy==2.4.1
folium==0.15.1
geopandas==0.14.2

# API Framework
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3

# Database
psycopg2-binary==2.9.9
redis==5.0.1
pymongo==4.6.1

# Machine Learning & AI
openai==1.10.0
langchain==0.1.0
chromadb==0.4.22  # Vector database
pinecone-client==3.0.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7
rich==13.7.0  # Better console output
schedule==1.2.0
celery==5.3.6  # Task queue

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0